Add core infrastructure to support Opportunistic Huge Pages
Opportunitistic Huge Page Support is aimed at selectively
using huge pages. A process must be chosen only if its
virtual-to-physical address translation overheads seem
to a be a potential performance bottleneck.

In the initial phase, we expect the userland to provide us
the list of processes to be considered for huge pages.

Signed-off-by: Panwar, Ashish <panwarit014@gmail.com>