OHP promotion based on the task weight
Each time we select a huge page candidate, we need to make
sure that the right task is being chosen. This patch selects
the next huge page candidate based on the weight of tasks
participating in the OHP list.

Right now, we traverse the list sequentially for selecting the
best candidate. A more efficient implementation should use rb
trees.

Signed-off-by: Panwar, Ashish <panwarit014@gmail.com>