Remove alloc sleep millisecs
We do not want to restrict the number of huge page allocations
per second as it can not adapt itself when the state of memory
fragmentation changes.

Rather, it makes more sense to restrict the average CPU utilization
of khugepaged. This patch uses 5% as the threshold and has been
observed to adjust the rate of promotion depending upon the state
of memory fragmentation.

Signed-off-by: Panwar, Ashish <panwarit014@gmail.com>